requirements.txt
flask==3.0.0
flask-login==0.6.3
flask-bcrypt==1.0.1
pymongo==4.6.1
python-dotenv==1.0.0
openpyxl==3.1.2
requests==2.31.0
gunicorn==21.2.0
dnspython==2.4.2
